+++ src/core/client/api/consts/cameraConst.ts
@@ -1,9 +1,11 @@
 import { CameraTarget } from '@AthenaClient/systems/cameraTarget';
+import { GameplayCamera } from '@AthenaClient/systems/gameplayCamera';
 import PedEditCamera from '@AthenaClient/utility/camera';
 import { CinematicCam } from '@AthenaClient/utility/cinematic';
 
 export const cameraConst = {
     ped: PedEditCamera,
+    gameplay: GameplayCamera,
     cinematic: CinematicCam,
     target: CameraTarget,
 };